results: update GMM
