+++ examples/gmm/evaluation_expectation_quadratic_func.py
@@ -6,6 +6,7 @@ import pandas as pd
 import os
 from omegaconf import DictConfig
 import torch
+import numpy as np
 
 from fab import FABModel, HamiltonianMonteCarlo, Metropolis
 from examples.make_flow import make_wrapped_normflowdist
@@ -49,7 +50,7 @@ def load_model(cfg: DictConfig, target, model_name: str):
 
 
 
-def evaluate(cfg: DictConfig, model_name: str, num_samples=int(1e4)):
+def evaluate(cfg: DictConfig, model_name: str, num_samples=int(1e3), n_repeats=100):
     torch.set_default_dtype(torch.float32)
     torch.manual_seed(cfg.training.seed)
     target = GMM(dim=cfg.target.dim, n_mixes=cfg.target.n_mixes,
@@ -58,16 +59,46 @@ def evaluate(cfg: DictConfig, model_name: str, num_samples=int(1e4)):
     if cfg.training.use_64_bit:
         torch.set_default_dtype(torch.float64)
         target = target.double()
-    model = load_model(cfg, target, model_name)
-    eval = model.get_eval_info(num_samples, 500)
-    return eval
+
+    biases = []
+    biases_unweighted = []
+    for i in range(n_repeats):
+        if model_name[:6] == "target":
+            samples = target.sample((num_samples, ))
+            log_w = torch.ones(samples.shape[0])
+            samples_unweighted = samples  # fake
+            log_w_unweighted = log_w  # fake
+        else:
+            model = load_model(cfg, target, model_name)
+            if model_name and model_name[0:3] == "snf":
+                samples, log_w = model.snf.sample(num_samples)
+                valid_indices_unweighted = ~ (torch.softmax(log_w, axis=0) == 0)
+                samples_unweighted = samples[valid_indices_unweighted]
+                log_w_unweighted = torch.ones_like(log_w[valid_indices_unweighted])
+            else:
+                samples, log_q = model.flow.sample_and_log_prob((num_samples, ))
+                log_w = target.log_prob(samples) - log_q
+                valid_indices = ~torch.isinf(log_w) & ~torch.isnan(log_w)
+                samples, log_w = samples[valid_indices], log_w[valid_indices]
+                valid_indices_unweighted = ~ (torch.softmax(log_w, axis=0) == 0)
+                samples_unweighted = samples[valid_indices_unweighted]
+                log_w_unweighted = torch.ones_like(log_w[valid_indices_unweighted])
+        normed_bias = target.evaluate_expectation(samples, log_w).detach().cpu()
+        normed_bias_unweighted = target.evaluate_expectation(samples_unweighted,
+                                                             log_w_unweighted).detach().cpu()
+        biases.append(normed_bias)
+        biases_unweighted.append(normed_bias_unweighted)
+    info = {"bias": np.mean(np.abs(biases)),
+            "std": np.std(biases),
+            "bias_unweighted": np.abs(np.mean(biases_unweighted))}
+    return info
 
 
-@hydra.main(config_path="../../config", config_name="gmm.yaml")
+@hydra.main(config_path="../config", config_name="gmm.yaml")
 def main(cfg: DictConfig):
-    model_names = ["fab_buffer", "fab_no_buffer", "flow_kld", "flow_nis", "target_kld", "snf"]
+    model_names = ["target", "fab_buffer", "fab_no_buffer", "flow_kld", "flow_nis", "target_kld", "snf"]
     seeds = [1, 2, 3]
-    num_samples = int(5e4)
+    num_samples = int(1000)
 
     results = pd.DataFrame()
     for model_name in model_names:
@@ -78,26 +109,24 @@ def main(cfg: DictConfig):
             eval_info.update(seed=seed,
                              model_name=model_name)
             results = results.append(eval_info, ignore_index=True)
+        print(results)
 
-    keys = ["eval_ess_flow", "eval_ess_ais", "test_set_mean_log_prob", 'kl_forward']
+    fields = ["bias", "bias_unweighted"]
     print("\n *******  mean  ********************** \n")
-    print(results.groupby("model_name").mean()[keys])
+    print(results.groupby("model_name").mean()[fields])
     print("\n ******* std ********************** \n")
-    print(results.groupby("model_name").sem(ddof=0)[keys])
-    print("overall results")
-    print(results[["model_name", "seed", "eval_ess_flow", "eval_ess_ais", "test_set_mean_log_prob"]])
-    results.to_csv(open(FILENAME_EVAL_INFO, "w"))
+    print(results.groupby("model_name").std()[fields])
+    results.to_csv(open(FILENAME_EXPECTATION_INFO, "w"))
 
-
-FILENAME_EVAL_INFO = "/home/laurence/work/code/FAB-TORCH/examples/paper_results/gmm/gmm_results.csv"
+FILENAME_EXPECTATION_INFO = "/examples/paper_results/gmm/gmm_results_expectation.csv"
 
 if __name__ == '__main__':
-    main()
-    # results = pd.read_csv(open("gmm_results.csv", "r"))
-    # print("mean")
-    # print(results.groupby("model_name").mean()[["eval_ess_flow", "eval_ess_ais", "test_set_mean_log_prob"]])
-    # print("std")
-    # print(results.groupby("model_name").std()[["eval_ess_flow", "eval_ess_ais", "test_set_mean_log_prob"]])
-    # results.to_csv(open("gmm_results.csv", "w"))
-    # print("overall results")
-    # print(results[["model_name", "seed", "eval_ess_flow", "eval_ess_ais", "test_set_mean_log_prob"]])
+    if True:
+        main()
+    else:
+        results = pd.read_csv(open(FILENAME, "r"))
+        fields = ["bias", "std"]
+        print("\n *******  mean  ********************** \n")
+        print(results.groupby("model_name").mean()[fields])
+        print("\n ******* std ********************** \n")
+        print(results.groupby("model_name").std()[fields])
