+++ examples/gmm/plots/MoG.png
