+++ examples/many_well/plots/many_well.png
