+++ examples/many_well/plots/rejection_sampling.png
