+++ examples/many_well/rejection_sampling_vis.py
