+++ examples/many_well/results_vis.py
@@ -1,12 +1,11 @@
 import os
-from typing import Optional
 import hydra
 import matplotlib.pyplot as plt
 from matplotlib import rc
 import matplotlib as mpl
 from omegaconf import DictConfig
 from examples.make_flow import make_wrapped_normflowdist
-from examples.many_well_visualise_all_marginal_pairs import get_target_log_prob_marginal_pair
+from examples.many_well.many_well_visualise_all_marginal_pairs import get_target_log_prob_marginal_pair
 from fab.utils.plotting import plot_contours, plot_marginal_pair
 from fab.target_distributions.many_well import ManyWellEnergy
 import torch
@@ -55,7 +54,7 @@ def plot_marginals(cfg: DictConfig, supfig, model_name, plot_y_label):
 
 
 
-@hydra.main(config_path="./", config_name="config.yaml")
+@hydra.main(config_path="/", config_name="config.yaml")
 def run(cfg: DictConfig):
     mpl.rcParams['figure.dpi'] = 300
     rc('font', **{'family': 'serif', 'serif': ['Times']})
