--- examples/paper_results/gmm/gmm_results.csv
@@ -1,19 +1,0 @@
-,eval_ess_flow,eval_ess_ais,test_set_mean_log_prob,bias_normed,bias_no_correction,ess_over_p,kl_forward,seed,model_name
-0,0.6046956911863202,0.6207629541543146,-7.21664690728586,0.08099836194171423,0.085083851419022,0.0011986705146966564,0.3631963879776271,1.0,fab_buffer
-1,0.683273185662027,0.6947995218157687,-7.113452636816453,0.08444411455284757,0.08510532269109965,0.0010568881314579992,0.26849387792303686,2.0,fab_buffer
-2,0.6338404611292385,0.6499832609136581,-7.188827171745133,0.08424762554739808,0.0648403509608125,1.64006124281932e-09,0.3300787314970768,3.0,fab_buffer
-3,0.29448316581593537,0.3090874247146333,-7.817212258860984,0.08632706762708509,0.11272474748361906,0.15699910637293124,0.9618848806628645,1.0,fab_no_buffer
-4,0.13366855176474168,0.13895088236179476,-8.119453377891183,0.11119956600840912,0.338748370672428,0.09296654381827392,1.2653557813398106,2.0,fab_no_buffer
-5,0.13914471474666973,0.14388042205886806,-8.058080964796014,0.084900027576882,0.06064526342758631,0.059485213436548626,1.2083003547106967,3.0,fab_no_buffer
-6,0.7322881998457096,0.753327502523118,,-0.20617817619980577,0.28790101833972026,,,1.0,flow_kld
-7,0.4968822337740068,0.38454560544929933,,-0.9241483289652694,0.8972387386052906,,,2.0,flow_kld
-8,0.9799158088434109,0.11375274032600331,,-0.9204117756872344,0.9228606239085496,,,3.0,flow_kld
-9,0.000393982998225116,0.00042604218949997524,,-0.9992921749903665,0.9984682409761027,,,1.0,flow_nis
-10,0.01729841680442923,0.008116565649342882,,-0.9980468416980265,0.9981353806765847,,,2.0,flow_nis
-11,2.313621127595143e-05,0.12078384798169141,,-1.03555036052916,1.0300146616419001,,,3.0,flow_nis
-12,0.61318572958185,0.6402183956201185,-7.155915732380058,0.08816871007811942,0.10980342373198625,2.670480541087992e-09,0.2935528719103725,1.0,target_kld
-13,0.6263060065525827,0.40547344365541166,-7.162951281251604,0.08328904905182304,0.07124256217347803,1.5265039047985995e-10,0.3011745686685666,2.0,target_kld
-14,0.4985875264493928,0.5362224908495891,-7.26861434815601,0.10368301380106455,0.05264106876452729,0.05794753927863472,0.4100240251088647,3.0,target_kld
-15,0.0679957914141388,,,0.9297863501095277,0.928867269375741,,,1.0,snf
-16,0.00010272818333270379,,,0.9015173693902523,0.9447426254589962,,,2.0,snf
-17,0.22867853611194924,,,0.9426509780874233,0.9404383668040354,,,3.0,snf
