--- examples/paper_results/gmm/gmm_results_expectation.csv
@@ -1,22 +1,0 @@
-,bias,std,bias_unweighted,seed,model_name
-0,0.08794610343472717,0.03688348538599228,0.08794610343472717,1.0,target
-1,0.09501647194417101,0.03876630984376682,0.09501647194417101,2.0,target
-2,0.08949530541713425,0.03417408924542846,0.08949530541713425,3.0,target
-3,0.0899981799524166,0.05271957721488359,0.08982638291847843,1.0,fab_buffer
-4,0.08409862618440327,0.051901464273671705,0.08357975678712437,2.0,fab_buffer
-5,0.08595762213906881,0.04990720417830446,0.07063246827467096,3.0,fab_buffer
-6,0.09783073994173933,0.07795624734240948,0.11719475693229703,1.0,fab_no_buffer
-7,0.14305525102285888,0.18609019036516575,0.39053237581480554,2.0,fab_no_buffer
-8,0.1000018183056675,0.08419082248156036,0.06589905014364456,3.0,fab_no_buffer
-9,0.20683123905188475,0.02562302068284853,0.2893849978694675,1.0,flow_kld
-10,0.9242288828710177,0.007505004817201208,0.8967975753991898,2.0,flow_kld
-11,0.9213100315005623,0.0022783819523784607,0.9230723287518602,3.0,flow_kld
-12,0.9990580103580592,0.00028602464245707914,0.9984524894700688,1.0,flow_nis
-13,0.9981287024485824,0.0003891821489728195,0.998149544702663,2.0,flow_nis
-14,1.0153674015595022,0.004436903637683524,1.0752819999358265,3.0,flow_nis
-15,0.0843251797192319,0.041377669776709855,0.08399806916667135,1.0,target_kld
-16,0.08610568232261091,0.04085668496610647,0.08557974684809064,2.0,target_kld
-17,0.07968296982438973,0.04183358776289322,0.07805900689869617,3.0,target_kld
-18,0.9267562329042839,0.004163189680004574,0.9288864121712457,1.0,snf
-19,0.9519688722257367,0.002186434416821341,0.9450591528796914,2.0,snf
-20,0.938268776083837,0.002297318411602757,0.9404912105200204,3.0,snf
